+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add support for HTTP_X_REAL_IP
 - adjust for MutableSet changes
 - add support for MySQL 8.4
+- ignore extraneous vscode triggered updates in watched directory
+- add support for mp4 responses
 
 ## [6.23.0] - 2024-04-24
 - add site timezone attribute